firebase storage call moved up into app level props not passed down to ActiveTasks
