+++ src/App.tsx
@@ -10,15 +10,41 @@ import Leaderboard from "./pages/Leaderboard/Leaderboard";
 import Profile from "./pages/Profile/Profile";
 import Register from "./pages/Register/Register";
 import Account from "./pages/Account/Account";
+import { app } from "./firebase";
+import { getFirestore } from "firebase/firestore";
+import { doc, getDoc } from "firebase/firestore";
+import { useEffect, useState } from "react";
+import { Task } from "./mockData/mockActiveTasks";
 
 const App = () => {
+  const [activeTasksList, setActiveTasksList] = useState<Task[]>([]);
+
+  useEffect(() => {
+    const getTasks = async () => {
+      const db = getFirestore(app);
+      const retrievalReference = doc(
+        db,
+        "test-active-tasks",
+        "OuZ1eeH9c5ZosgoXUi6Iraq7oM03"
+      );
+      const retrieveTasks = await getDoc(retrievalReference);
+      if (retrieveTasks.exists()) {
+        console.log("Document data:", retrieveTasks.data());
+        setActiveTasksList(retrieveTasks.data().activeTasks);
+      } else {
+        console.log("No such document!");
+      }
+    };
+    getTasks();
+  }, []);
+
   return (
     <>
       <Routes>
         <Route path="/" element={<Home />} />
         <Route
           path="/tasks"
-          element={<ActiveTasks userId={"qDjHyzko7ehZKSOSHe0uHJ0KEjR2"} />}
+          element={<ActiveTasks activeTasks={activeTasksList} />}
         />
         <Route
           path="/calendar"
